Update mobilenet.js
Attempting to run this file I was getting an error in node `(node:14520) UnhandledPromiseRejectionWarning: TypeError: Cannot read property 'length' of undefined` it seems that `out_data` is looking for data in a spot that it doesn't exist. So I changed it to just look in the last position and it works.